# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "user login",
                "description": "A user enters login details",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "required|string"
                                    },
                                    "password": {
                                        "type": "required|string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "login successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "login fails",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/project-create": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "create Project",
                "description": "Admin create a project",
                "operationId": "f37e982c458c627efef63f2388aa74e4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Project_name"
                                ],
                                "properties": {
                                    "Project_name": {
                                        "type": "required|string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "fails",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "api/project-create": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get all Projects",
                "description": "Get all Projects",
                "operationId": "b039e2f218054dd0a5615b5127feb85e",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "api/user-projects/{$project_id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get each Projects",
                "description": "Get each Projects",
                "operationId": "2c935d0d7286e2fbbdf6fd984929acb6",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/task-create": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "create tasks for a project",
                "description": "Admin create tasks for a project",
                "operationId": "798debdf2b924bcf207eca23a0bc8a7c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "task_name"
                                ],
                                "properties": {
                                    "task_name": {
                                        "type": "required|string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "fails",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "api/all-tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get all Tasks",
                "description": "Get all Tasks",
                "operationId": "19a6cfe944e20e381784f51f6a29b88f",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "api/user-tasks/{$task_id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get each Tasks",
                "description": "Get each Tasks",
                "operationId": "be63d483592299f1bbd9634e0bb19d72",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    }
}